Having OIDC config service extending OAuth2 one to keep it in sync with the related spec efforts
